+++ week5/src/main/java/com/example/demo1/Common/exception/type/ErrorType.java
@@ -1,4 +1,4 @@
-package com.example.demo1.Common.type;
+package com.example.demo1.Common.exception.type;
 
 public interface ErrorType {
     String getCode();
