+++ week5/src/main/java/com/example/demo1/Common/response/user/UserSuccessType.java
@@ -1,14 +1,14 @@
-package com.example.demo1.Common.type;
+package com.example.demo1.Common.response.user;
 
-public enum CommonErrorType implements ErrorType {
+import com.example.demo1.Common.response.type.SuccessType;
 
-    INTERNAL_SERVER("Common1", "Internal Server Error"),
-    BAD_REQUEST("Common2", "잘못된 요청입니다."),;
+public enum UserSuccessType implements SuccessType {
+    CREATE("USER_1", "유저 생성에 성공했습니다.");
 
     private final String code;
     private final String message;
 
-    CommonErrorType(String code, String message) {
+    UserSuccessType(String code, String message) {
         this.code = code;
         this.message = message;
     }
